moved remaining "option store" tests to test_runner.py
